Branched from $/epiinfo/Epi Info 7.1.4.0 Release to create the first check-in of Web Enter integration (9/30 Ebola Liberia)

+++ StatisticsRepository/EICSField.vb
@@ -38,11 +38,11 @@ Public Class CSField
     Private cbMissing As Boolean
     Private cenumFieldType As Integer
 
-    Public Function FieldInt() As Long
+    Public Function FieldInt() As Short
         ' return the integer value of the field
 
         If (IsNumeric(csFieldEntry) And CDec(Val(csFieldEntry)) Mod 1 = 0) Then
-            FieldInt = CLng(Val(csFieldEntry))
+            FieldInt = CInt(Val(csFieldEntry))
         Else
             FieldInt = 0.0#
         End If
